+++ src/behaviors.lua
@@ -1,5 +1,4 @@
 local behaviors = {
-    patches = {},
     hooks = {},
     options = {},
 }
@@ -16,15 +15,12 @@ end
 local function register(path)
     local behavior = import(path)
     appendOptions(behavior)
-    for _, patch in ipairs(behavior.patches or {}) do
-        table.insert(behaviors.patches, patch)
-    end
     for _, hook in ipairs(behavior.hooks or {}) do
         table.insert(behaviors.hooks, hook)
     end
 end
 
-register("behaviors/UnlimitedBoonSkips.lua")
+register("behaviors/VowRewardSkip.lua")
 register("behaviors/SpecialBoonFallbacks.lua")
 
 return behaviors